Fix: userId int 통일, 게임 종료 로직 수정

+++ holyshit-server/Src/Data/Room.cs
@@ -8,12 +8,12 @@ public class Room
   public int Id { get; set; }
   public string Name { get; set; } = string.Empty;
   public int MaxUserNum { get; set; }
-  public long OwnerId { get; set; }
+  public int OwnerId { get; set; }
   public RoomStateType State { get; set; }
 
   // 동시성을 고려한 컬렉션들
-  private readonly ConcurrentDictionary<long, UserData> _users = new();
-  private readonly ConcurrentDictionary<long, CharacterPositionData> _characterPositions = new();
+  private readonly ConcurrentDictionary<int, UserData> _users = new();
+  private readonly ConcurrentDictionary<int, CharacterPositionData> _characterPositions = new();
 
   // 유저 관리 메서드들
   public bool AddUser(UserData userData)
@@ -21,12 +21,12 @@ public class Room
     return _users.TryAdd(userData.Id, userData);
   }
 
-  public bool RemoveUser(long userId)
+  public bool RemoveUser(int userId)
   {
     return _users.TryRemove(userId, out _);
   }
 
-  public UserData? GetUser(long userId)
+  public UserData? GetUser(int userId)
   {
     _users.TryGetValue(userId, out var userData);
     return userData;
@@ -53,7 +53,7 @@ public class Room
   }
 
   // 위치 관리 메서드들
-  public bool UpdatePosition(long userId, double x, double y)
+  public bool UpdatePosition(int userId, double x, double y)
   {
     try
     {
@@ -92,7 +92,7 @@ public class Room
       OwnerId = OwnerId,
       State = State
     };
-    
+
     roomData.Users.AddRange(GetAllUsers());
     return roomData;
   }
